Changes
* dont exceed byte size for token decimals
* move TokenSummary to within CryptoNote
* TODO delete IToken.h
* TODO fix name arg in conceal wallet
* calculate token amounts with correct decimals rather than $CCX decimals
*

+++ src/CryptoNoteCore/Blockchain.h
@@ -236,7 +236,7 @@ namespace cn
 
       uint64_t token_id;
       uint64_t token_supply;
-      uint64_t decimals;
+      uint8_t decimals;
       uint64_t created_height;
       std::string ticker;
       std::string token_name;
